Removed flaky test reporting upstream
https://github.com/envoyproxy/envoy/commit/92f1cd4685191ef26409230999a40fbca96f5247

+++ ci/do_ci.sh
@@ -13,8 +13,13 @@ cp -f $UPSTREAM_ENVOY_SRCDIR/.bazelversion $SOURCE_DIR/.bazelversion
 # cp -f $UPSTREAM_ENVOY_SRCDIR/bazel/get_workspace_status $SOURCE_DIR/bazel/get_workspace_status
 cp -f $UPSTREAM_ENVOY_SRCDIR/ci/WORKSPACE.filter.example $SOURCE_DIR/ci/
 
+# upstream removed the flaky_test
 mkdir -p $SOURCE_DIR/ci/flaky_test
-cp -a $UPSTREAM_ENVOY_SRCDIR/ci/flaky_test $SOURCE_DIR/ci
+# Shooting in the dark here since I cannot view logs. These were removed upstream
+# but the build seems to fail here.
+if [[ -d "$UPSTREAM_ENVOY_SRCDIR/ci/flaky_test" ]] || [[ -f "$UPSTREAM_ENVOY_SRCDIR/ci/flaky_test" ]]; then
+  cp -a $UPSTREAM_ENVOY_SRCDIR/ci/flaky_test $SOURCE_DIR/ci
+fi
 
 cp -f $UPSTREAM_ENVOY_SRCDIR/tools/shell_utils.sh $SOURCE_DIR/tools
